Add button to add channels to dashboard
