+++ src/app/api/feed/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { PrismaClient } from "@prisma/client";
 
-import { prisma } from "../../../lib/prisma";
+import prisma from "../../../lib/prisma";
 
 export async function GET() {
   try {
