+++ src/lib/prisma.ts
@@ -1,7 +1,7 @@
 import { PrismaClient } from "@prisma/client";
 
-const globalForPrisma = global as unknown as { prisma?: PrismaClient };
+// Initialize the Prisma Client
+const prisma = new PrismaClient();
 
-export const prisma = globalForPrisma.prisma ?? new PrismaClient();
-
-if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
+// Export the Prisma Client instance
+export default prisma;
